Fix bug 3141636 	documentNumber limited to 2^32 for 9.1, 9.2, and 9.3 Fix bug 3140874 	AIOOB in TinyTree.addNamespace() for 9.1

+++ latest9.1/bj/net/sf/saxon/dom/DocumentWrapper.java
@@ -19,7 +19,7 @@ public class DocumentWrapper extends NodeWrapper implements DocumentInfo {
 
     protected Configuration config;
     protected String baseURI;
-    protected int documentNumber;
+    protected long documentNumber;
     protected boolean domLevel3;
 
     /**
@@ -93,7 +93,7 @@ public class DocumentWrapper extends NodeWrapper implements DocumentInfo {
      * Get the unique document number
      */
 
-    public int getDocumentNumber() {
+    public long getDocumentNumber() {
         return documentNumber;
     }
