+++ latest9.1/bj/net/sf/saxon/jdom/DocumentWrapper.java
@@ -21,7 +21,7 @@ public class DocumentWrapper extends NodeWrapper implements DocumentInfo {
 
     protected Configuration config;
     protected String baseURI;
-    protected int documentNumber;
+    protected long documentNumber;
 
     /**
      * Create a Saxon wrapper for a JDOM document
@@ -85,7 +85,7 @@ public class DocumentWrapper extends NodeWrapper implements DocumentInfo {
 	* Get the unique document number
 	*/
 
-	public int getDocumentNumber() {
+	public long getDocumentNumber() {
 	    return documentNumber;
 	}
